tools/orchai2/requirements.txt
docker>=6.1.3
pyyaml>=6.0.1
python-dotenv>=1.0.0
rocketchat-API>=1.30.0
requests>=2.31.0